backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.101.0
filelock==3.18.0
Flask==3.0.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.13.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
Pillow==10.0.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.21
pydantic_core==2.33.1
Pygments==2.19.1
python-multipart==0.0.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
torch==2.6.0
torchvision==0.21.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.23.2
Werkzeug==3.1.3
wrapt==1.14.1
